@@ -1,3 +1,3 @@
 [tool.black]
 line-length = 110
-target-version = ['py35', 'py36', 'py37', 'py38']
+target-version = ['py36', 'py37', 'py38']
